# Flash slot to key