--- a/src/storage/rocksdb.h
+++ b/src/storage/rocksdb.h
@@ -32,6 +32,7 @@ public:
     virtual void retrieve(const char *key, size_t cchKey, callbackSingle fn) const override;
     virtual size_t clear() override;
     virtual bool enumerate(callback fn) const override;
+    virtual bool enumerate_hashslot(callback fn, int hashslot) const override;
 
     virtual const IStorage *clone() const override;