--- a/src/storage/teststorageprovider.h
+++ b/src/storage/teststorageprovider.h
@@ -24,6 +24,7 @@ public:
     virtual void retrieve(const char *key, size_t cchKey, callbackSingle fn) const override;
     virtual size_t clear() override;
     virtual bool enumerate(callback fn) const override;
+    virtual bool enumerate_hashslot(callback fn, int hashslot) const override;
     virtual size_t count() const override;
 
     virtual void flush() override;