This change replaces to the slot_to_key map when flash in enabled by prefixing each key with its hashslot when stored in rocksdb and using rocksdb prefix iter seek to replace functionality of the slot to key map.

Also add --flash option to test scripts to run cluster tests with flash.

maxmemory 100MB, `memtier_benchmark -n 1000000 --cluster`
with change:
```
ALL STATS
======================================================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    MOVED/sec      ASK/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
------------------------------------------------------------------------------------------------------------------------------------------------------
Sets        20330.84          ---          ---         0.00         0.00         2.71533         2.31900         8.70300        14.07900      1566.28 
Gets       203306.19    192656.60     10649.59         0.00         0.00         2.65851         2.27100         8.63900        14.01500     14316.36 
Waits           0.00          ---          ---          ---          ---             ---             ---             ---             ---          --- 
Totals     223637.03    192656.60     10649.59         0.00         0.00         2.66367         2.27100         8.63900        14.01500     15882.64 
```
without change:
```
ALL STATS
======================================================================================================================================================
Type         Ops/sec     Hits/sec   Misses/sec    MOVED/sec      ASK/sec    Avg. Latency     p50 Latency     p99 Latency   p99.9 Latency       KB/sec 
------------------------------------------------------------------------------------------------------------------------------------------------------
Sets         5428.89          ---          ---         0.00         0.00         9.15187         2.01500       202.75100       313.34300       418.24 
Gets        54288.30     51452.47      2835.83         0.00         0.00         8.99135         1.96700       200.70300       313.34300      3823.12 
Waits           0.00          ---          ---          ---          ---             ---             ---             ---             ---          --- 
Totals      59717.19     51452.47      2835.83         0.00         0.00         9.00595         1.96700       200.70300       313.34300      4241.36 
```
Fixes #646.