@JohnSully 2023-05-11
Waiting on:
- memtier results for read/write
- Configure rocksdb to only consider the slot prefix